requirements.txt
certifi==2024.7.4
cffi==1.17.1
chardet==3.0.4
charset-normalizer==2.0.10
cryptography==43.0.1
decorator==4.4.1
future==0.18.3
idna==3.7
pycparser==2.19
python-telegram-bot==12.4.2
requests==2.32.0
six==1.14.0
tornado==6.4.2
urllib3==2.2.2
secret~=0.8